Frontend is talking to Backend :)
Next I'm going to make sure the form submission is working correctly, and that I can add the value in the form to the DB. Then I'll work on the second page so the user can see their submitted expenses. As an aside, if I have time I should probably implement some kind of sign-in feature, since it's likely there would be more than one user for this application.